Készlet és üresen megnyitás

+++ src/main/java/com/raktar3/repository/CompanyRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.raktar3.entities.Company;
 
-public interface CompanyRepository extends CrudRepository<Company, Long> {
+public interface CompanyRepository extends CrudRepository<Company, Integer> {
 
 	
 	List<Company> findAll();
